move model.ts
