+++ skymp5-client/src/gamemodeApi/gamemodeApiSupport.ts
@@ -1,6 +1,6 @@
 import { ObjectReference, on, printConsole, storage } from "skyrimPlatform";
 import * as sp from "skyrimPlatform";
-import { FormModel } from "../modelSource/model";
+import { FormModel } from "../view/model";
 import { FormViewArray } from "../view/formViewArray";
 import { localIdToRemoteId, remoteIdToLocalId } from "../view/worldViewMisc";
