+++ skymp5-client/src/sync/movementGet.ts
@@ -1,4 +1,4 @@
-import { FormModel } from '../modelSource/model';
+import { FormModel } from '../view/model';
 import { ObjectReference, Actor, TESModPlatform } from "skyrimPlatform";
 import { NiPoint3, Movement, RunMode } from "./movement";
 import { ObjectReferenceEx } from '../extensions/objectReferenceEx';
