+++ skymp5-client/src/view/formViewArray.ts
@@ -1,6 +1,6 @@
 import { FormView } from "./formView";
 import { GamemodeApiSupport } from "../gamemodeApi/gamemodeApiSupport";
-import { FormModel, WorldModel } from "../modelSource/model";
+import { FormModel, WorldModel } from "./model";
 import { Movement, NiPoint3 } from "../sync/movement";
 
 export class FormViewArray {
