+++ skymp5-client/src/view/model.ts
